backend/requirements.txt
fastapi>=0.111,<1.0
uvicorn[standard]>=0.30,<1.0
pydantic-settings>=2.3,<3.0
SQLAlchemy>=2.0.31,<3.0
alembic>=1.18.4,<2.0
PyJWT>=2.8.0,<3.0
pypdf>=6.10.2,<7.0
email-validator>=2.2,<3.0
python-multipart>=0.0.9,<1.0
pytest>=8.2,<9.0
httpx>=0.27,<1.0
